+++ corvis/src/feature/descriptor/patch_descriptor.h
@@ -27,6 +27,6 @@ public:
 
     patch_descriptor(float x, float y, const tracker::image& image);
     static bool is_better(const float score1, const float score2) {return score1 > score2;}
-    static double distance(const patch_descriptor &a, const patch_descriptor &b);
-    double distance(float x, float y, const tracker::image& image) const;
+    static float distance(const patch_descriptor &a, const patch_descriptor &b);
+    float distance(float x, float y, const tracker::image& image) const;
 };
